requirements.txt
PyCrypto
requests
